+++ view/statistics-time-per-person.js
@@ -16,6 +16,7 @@ var copy                 = require('es5-ext/object/copy')
   , getDurationDaysHours = require('./utils/get-duration-days-hours-fine-grain')
   , dateFromToBlock      = require('./components/filter-bar/select-date-range-safe-fallback')
   , getDynamicUrl        = require('./utils/get-dynamic-url')
+  , statisticsTimeRowOnClick        = require('./utils/statistics-time-row-onclick')
   , processingStepsMetaWithoutFrontDesk
 	= require('./utils/processing-steps-meta-without-front-desk');
 
@@ -43,7 +44,7 @@ var getRowResult = function (rowData, label) {
 
 exports['statistics-main'] = function () {
 	var stepsMeta = processingStepsMetaWithoutFrontDesk(),
-		stepsMap = {}, queryHandler, params;
+		stepsMap = {}, queryHandler, params, queryResult;
 	Object.keys(stepsMeta).forEach(function (stepShortPath) {
 		stepsMap[stepShortPath]   = new ObservableValue();
 	});
@@ -61,18 +62,22 @@ exports['statistics-main'] = function () {
 			query.dateTo = query.dateTo.toJSON();
 		}
 		queryServer(query).done(function (result) {
+			queryResult = result;
 			Object.keys(stepsMap).forEach(function (key) {
-				var preparedResult = [];
+				var preparedResults = [];
 				if (!result.byStep[key]) {
 					stepsMap[key].value = null;
 					return;
 				}
 				forEach(result.byStepAndProcessor[key], function (rowData, userId) {
-					preparedResult.push(getRowResult(rowData.processing,
-						db.User.getById(userId).fullName));
+					var preparedResult = getRowResult(rowData.processing,
+						db.User.getById(userId).fullName);
+					preparedResult.key = key;
+					preparedResult.userId = userId;
+					preparedResults.push(preparedResult);
 				});
-				preparedResult.push(getRowResult(result.byStep[key].processing, _("Total & times")));
-				stepsMap[key].value = preparedResult;
+				preparedResults.push(getRowResult(result.byStep[key].processing, _("Total & times")));
+				stepsMap[key].value = preparedResults;
 			});
 		});
 	});
@@ -140,13 +145,27 @@ exports['statistics-main'] = function () {
 							onEmpty: tr(td({ class: 'empty statistics-table-number', colspan: 5 },
 								_("There are no files processed at this step")))
 						}, data, function (rowData) {
-							return tr(
+
+							var step, props = {};
+
+							if (rowData && rowData.key && rowData.userId) {
+								step = queryResult.byStepAndProcessor[rowData.key][rowData.userId];
+							}
+
+							if (step && step.businessProcesses.length !== 0) {
+								props.class = 'cursor-pointer text-decoration-underline';
+								props.onclick = function () {
+									statisticsTimeRowOnClick(window.jQuery(this), step.businessProcesses);
+								};
+							}
+
+							return tr(props,
 								td(rowData.label),
 								td({ class: 'statistics-table-number' }, rowData.timedCount),
 								td({ class: 'statistics-table-number' }, rowData.avgTime),
 								td({ class: 'statistics-table-number' }, rowData.minTime),
 								td({ class: 'statistics-table-number' }, rowData.maxTime)
-							);
+								);
 						})
 						)), br()];
 			});
